# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aryan Landmark Solutions API
  version: 1.0.0
  description: |
    Production REST API for Aryan Landmark Solutions.

    PostgreSQL contains six tables. All property photos, property documents
    and inquiry call recordings use one unified `attachments` resource.

    Cloudflare R2 stores file binaries. PostgreSQL stores metadata and the R2
    object key.
servers:
  - url: /api/v1
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      security:
        - applicationBearer: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: search
          in: query
          description: Matches against name, mobile, or email
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, BLOCKED]
      responses:
        '200':
          description: Customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'

    post:
      tags: [Customers]
      summary: Create customer
      description: |
        Creates a customer master record (users.user_type = CUSTOMER).
        Available to ADMIN and EMPLOYEE application users. createdBy is
        derived from the authenticated JWT, never accepted from the
        request body.
      security:
        - applicationBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer details
      security:
        - applicationBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Customers]
      summary: Delete customer
      description: |
        Phase 18A. ADMIN only — EMPLOYEE receives 403. Blocked with 409
        if any inquiry references this customer
        (inquiries.customer_id is ON DELETE RESTRICT at the database
        level — this is not a new rule, it predates lightweight
        inquiries). A customer with no inquiries can always be deleted.
      security:
        - applicationBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Customer deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: One or more inquiries reference this customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    applicationBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT

    masterBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Authorization failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Business conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateCustomerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        mobile:
          type: string
        alternateMobile:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
    Customer:
      type: object
      description: |
        Matches the actual response of CustomersService/customer.mapper.ts
        exactly (a distinct shape from User — passwordHash, userId, and
        role never apply to a CUSTOMER and are not part of this schema).
      required: [id, userType, name, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userType:
          type: string
          enum: [CUSTOMER]
        name:
          type: string
        email:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        alternateMobile:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        pincode:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, INACTIVE, BLOCKED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Customer'
    CustomerListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
        requestId:
          type: string
...